Prototype knowledge
What is a prototype?
A prototype is an object that acts as a template for another object. We have no classes in
JavaScript, so features are shared between objects through other objects taken as reference.
In other words, if we need an object that is similar to an existing object A, we create a new
object B saying that its prototype is the existing object A. This is the basic mechanism of
JavaScript's inheritance.